Update files message and file
